atividades/atv_05/backend/django/pyproject.toml
[project]
name = "hotel"
version = "0"
description = "Default template for PDM package"
authors = [
    {name = "davi-prussek", email = "davi.mariano.prussek0916@gmail.com"},
]
dependencies = ["django>=4.2.27", "mysqlclient>=2.2.7", "python-decouple>=3.8", "djangorestframework>=3.16.1"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
